addendum
I like Punta Cana a lot. The weather here is gorgeous, the people are fun, the food rocks, I've found a place where I feel welcome and happy. This airport is no different mostly. I mean, it's an airport, it's the same thing whether you're in Bangladesh or Sydney. And it always sucks. Worldwide suck. At least have the chairs have a little padding for fucks sake. Like fill them with sand or something. It's a beautiful building, and being up in the air really is a miracle. But this ain't it, you know? Everybody here works so hard, back breaking 14 hours shifts and then some. And it doesn't have to be this way. Meanwhile the dogs are barking if you've got a fantastic in your purse. And duty free means nothing to me. It's still 70 for a whiskey they upcharge that shit all day. I spent 14 on coffee, 24 on a Coors light and wine, I'm down a hundred. and if anything, this experience, this place, this is Nazi Germany if y'all gonna call everything Nazis. this bullshit is Nazis. Scan me at the front, scan me at the gate, make this whole deal a mall. You don't gotta treat everybody, spending an egregious amount of money on coffee, you don't gotta make us all feel like bin fucking laden.